fix: Complete the division between declaration and definitions in h / cpp files.
